prepare 3.2.0 release (#46)
## [3.2.0] - 2024-05-01
### Added:
- Added an optional timeout to the `waitForInitialization` method. When
a timeout is specified the returned promise will be rejected after the
timeout elapses if the client has not finished initializing within that
time. When no timeout is specified the returned promise will not be
resolved or rejected until the initialization either completes or fails.

### Changed:
- The track method now validates that the provided metricValue is a
number. If a metric value is provided, and it is not a number, then a
warning will be logged.

### Fixed:
- Fixed the documentation for `evaluationReasons` for the `identify`
method.
- Upgraded to launchdarkly-eventsource version 2.0.2. This version
contains performance improvements and logging improvements compared the
the version that was in use.

---------

Co-authored-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>
Co-authored-by: Eli Bishop <eli@launchdarkly.com>
Co-authored-by: LaunchDarklyReleaseBot <launchdarklyreleasebot@launchdarkly.com>
Co-authored-by: Louis Chan <lchan@launchdarkly.com>
Co-authored-by: Ember Stevens <ember.stevens@launchdarkly.com>
Co-authored-by: Ember Stevens <79482775+ember-stevens@users.noreply.github.com>
Co-authored-by: ld-repository-standards[bot] <113625520+ld-repository-standards[bot]@users.noreply.github.com>
Co-authored-by: Kane Parkinson <93555788+kparkinson-ld@users.noreply.github.com>
Co-authored-by: Matthew M. Keeler <mkeeler@launchdarkly.com>